Initial Upload of Previous Development Files
Initial upload of files from previous programming collaboration with
permission to continue development. Current focus of the project will be
in optimizing the game engine to allow for a larger scale game while
maintaining playability on a wide range of potential devices.

Large credits to Derek Paschal for developing the current code with me
over the course of the previous project! Much of the current program was
made possible by your insights and hard work.